+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.4.0
 
 - Change `StatefulForm.valueOf()` to return only non-null values.
-- Add `StatefulForm.valueOrNull()` method which allows `null` values.
+- Add `StatefulForm.valueOfNullable()` method which allows `null` values.
 
 ## 0.3.1